% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{changepoint_function}
\alias{changepoint_function}
\title{Helper function for the change point estimation}
\usage{
changepoint_function(sample, dist_to_margin = 1)
}
\arguments{
\item{sample}{Numeric matrix or numeric data frame (rows = n, cols = p).}

\item{dist_to_margin}{Integer index (1-based). Bounds the domain
of this function away from \code{1} and \code{nrow(sample)}, where uncontrollable
spikes of the function values might occur.}
}
\description{
This function will be used to calculate and/or plot the
change point function.
}
\seealso{
\code{\link[=estimate_changepoint]{estimate_changepoint()}}, \code{\link[=plot_changepoint_function]{plot_changepoint_function()}}
}
