% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence-interval.R
\name{ci_sq_norm_twoside}
\alias{ci_sq_norm_twoside}
\title{Two-sided (1 - \code{alpha}) confidence interval for
the (squared) normalized norm}
\usage{
ci_sq_norm_twoside(test_result, alpha = 0.05)
}
\arguments{
\item{test_result}{data frame that can be obtained by applying
\code{\link[=cp_test_normalized]{cp_test_normalized()}} or \code{\link[=cp_test_sparsity_adj]{cp_test_sparsity_adj()}}}

\item{alpha}{Confidence level in (0,1). Default 0.05.}
}
\value{
Named numeric vector: c(lower, upper)
}
\description{
Two-sided (1 - \code{alpha}) confidence interval for
the (squared) normalized norm
}
