% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{draw_h}
\alias{draw_h}
\title{Draw i.i.d. samples from the distribution \eqn{H_\alpha}}
\usage{
draw_h(n = 1, n_points = 20, alpha = 4)
}
\arguments{
\item{n}{Integer number of draws (>= 1).}

\item{n_points}{Number of support points of the probability
measure nu.}

\item{alpha}{Exponent in the integral, see manuscript.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Draw i.i.d. samples from the distribution \eqn{H_\alpha}
}
\examples{
draw_h(5, alpha = 4)

}
