% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{estimate_changepoint}
\alias{estimate_changepoint}
\title{Estimation of k hat (Change point location)}
\usage{
estimate_changepoint(sample, dist_to_margin = 1)
}
\arguments{
\item{sample}{Numeric matrix or numeric data frame (rows = n, cols = p).}

\item{dist_to_margin}{Integer index (1-based). Bounds the possible
values for khat away from \code{1} and \code{nrow(sample)}, where uncontrollable
spikes might occur.}
}
\value{
Numeric khat: a value between \code{1 + dist_to_margin} and
\code{nrow(sample) - dist_to_margin}.
}
\description{
Computes the estimator for the change point as described
in the manuscript.
}
\examples{
X <- matrix(rnorm(200 * 20), 200, 20)
khat <- estimate_changepoint(X)
khat_safe <- estimate_changepoint(X, 5)

}
\seealso{
\code{\link[=plot_changepoint_function]{plot_changepoint_function()}}, \code{\link[=changepoint_function]{changepoint_function()}}
}
