% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{plot_changepoint_function}
\alias{plot_changepoint_function}
\title{Estimation of k hat (Change point location)}
\usage{
plot_changepoint_function(sample)
}
\arguments{
\item{sample}{Numeric matrix or numeric data frame (rows = n, cols = p).}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Computes the estimator for the change point as described
in the manuscript.
}
\examples{
X <- matrix(rnorm(200 * 20), 200, 20)
plot_changepoint_function(X)

}
\seealso{
\code{\link[=changepoint_function]{changepoint_function()}}
}
